--- docs/resources/boot_resources.md
@@ -1,54 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "maas_boot_resources Resource - terraform-provider-maas"
-subcategory: ""
-description: |-
-  Provides a resource to manage MAAS boot resources.
----
-
-# maas_boot_resources (Resource)
-
-Provides a resource to manage MAAS boot resources.
-
-## Example Usage
-
-```terraform
-resource "maas_boot_source" "test_boot_source" {
-  url = "http://images.maas.io/ephemeral-v3/candidate/"
-}
-
-resource "maas_boot_source_selection" "jammy_test" {
-  boot_source = maas_boot_source.test_boot_source.id
-
-  os      = "ubuntu"
-  release = "jammy"
-}
-
-resource "maas_boot_source_selection" "noble_test" {
-  boot_source = maas_boot_source.test_boot_source.id
-
-  os      = "ubuntu"
-  release = "noble"
-}
-
-resource "maas_boot_resources" "test" {
-  boot_source = maas_boot_source.test_boot_source.id
-
-  boot_source_selections = [
-    maas_boot_source_selection.jammy_test.id,
-    maas_boot_source_selection.noble_test.id,
-  ]
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `boot_source_selections` (Set of Number) The set of database IDs for boot source selections to attach to this boot resource
-
-### Read-Only
-
-- `boot_source` (Number) The boot source database ID this resource set is associated with.
-- `id` (String) The ID of this resource.
