+++ maas/data_source_maas_boot_resources.go
@@ -3,9 +3,14 @@ package maas
 import (
 	"context"
 	"fmt"
+	"time"
 
+	"github.com/canonical/gomaasclient/client"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+	"github.com/canonical/gomaasclient/entity"
 )
 
 func dataSourceMAASBootResources() *schema.Resource {
@@ -91,4 +96,47 @@ func dataSourceMAASBootResourcesRead(ctx context.Context, d *schema.ResourceData
 	}
 
 	return nil
+}
+
+func getBootResources(client *client.Client, syncType string) ([]entity.BootResource, error) {
+	// syncType: one of synched, uploaded
+	readParams := entity.BootResourcesReadParams{
+		Type: syncType,
+	}
+
+	bootResources, err := client.BootResources.Get(&readParams)
+	if err != nil {
+		return nil, err
+	}
+
+	return bootResources, nil
+}
+
+func awaitImportComplete(client *client.Client) error {
+	if err := client.BootResources.Import(); err != nil {
+		return err
+	}
+
+	timeout := 40 * time.Minute
+
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+
+	result := retry.RetryContext(ctx, timeout, func() *retry.RetryError {
+		if importing, err := client.BootResources.IsImporting(); err != nil {
+			return retry.NonRetryableError(err)
+		} else if importing {
+			return retry.RetryableError(fmt.Errorf("boot resources still importing, waiting... "))
+		}
+
+		return nil
+	})
+	// add a small delay to ensure the resources are fully updated
+	if err := retry.RetryContext(ctx, 10*time.Second, func() *retry.RetryError {
+		return nil
+	}); err != nil {
+		return fmt.Errorf("error after waiting 10 seconds: %s", err)
+	}
+
+	return result
 }
