+++ maas/data_source_maas_boot_resources_test.go
@@ -8,6 +8,7 @@ import (
 	"terraform-provider-maas/maas/testutils"
 	"testing"
 
+	"github.com/canonical/gomaasclient/client"
 	"github.com/canonical/gomaasclient/entity"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
@@ -133,4 +134,19 @@ func testAccCheckDataSourceMAASBootResourcesDestroy(s *terraform.State) error {
 	}
 
 	return nil
+}
+
+func findBootSourceSelection(client *client.Client, bootSource int, os string, release string) (*entity.BootSourceSelection, error) {
+	bootSourceSelections, err := client.BootSourceSelections.Get(bootSource)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, d := range bootSourceSelections {
+		if d.OS == os && d.Release == release {
+			return &d, nil
+		}
+	}
+
+	return nil, err
 }
